+++ R/dtstm.R
@@ -153,7 +153,8 @@ CohortDtstmTrans <- R6::R6Class("CohortDtstmTrans",
 #' @param trans_mat A transition matrix describing the states and transitions 
 #' in a discrete-time multi-state model. See [CohortDtstmTrans].
 #' @param n Number of random observations of the parameters to draw.
-#' @param point_estimate If `TRUE`, then the point estimates are returned and and no samples are drawn.
+#' @param uncertainty Method determining how parameter uncertainty should be handled. See
+#'  documentation in [`create_params()`].
 #' @export
 create_CohortDtstmTrans <- function(object, ...){
   UseMethod("create_CohortDtstmTrans", object)
@@ -163,12 +164,24 @@ create_CohortDtstmTrans <- function(object, ...){
 #' @rdname create_CohortDtstmTrans
 create_CohortDtstmTrans.multinom_list <- function(object, input_data,
                                                   trans_mat,
-                                                  n = 1000, point_estimate = FALSE,
+                                                  n = 1000, 
+                                                  uncertainty = c("normal", "none"),
                                                   ...){
+  # For backwards compatibility until deprecated point_estimate argument is no longer supported
+  dots <- list(...)  
+  uncertainty <- deprecate_point_estimate(dots$point_estimate, uncertainty,
+                                          missing(uncertainty))
+  dots <- dots[names(dots) != "point_estimate"]
+  
+  # Code to always keep
+  uncertainty <- match.arg(uncertainty)
   input_mats <- create_input_mats(object, input_data)
-  params <- create_params(object, n = n, point_estimate = point_estimate)
-  return(CohortDtstmTrans$new(params = params, input_data = input_mats, 
-                              trans_mat = trans_mat, ...))
+  params <- create_params(object, n = n, uncertainty = uncertainty)
+  return(
+    do.call(CohortDtstmTrans$new, 
+            c(list(params = params, input_data = input_mats, trans_mat = trans_mat),
+              dots))
+  )
 }
 
 # CohortDtstm ------------------------------------------------------------------
@@ -229,8 +242,7 @@ create_CohortDtstmTrans.multinom_list <- function(object, input_data,
 #' transmod <- create_CohortDtstmTrans(transfits,
 #'                                     input_data = transmod_data,
 #'                                     trans_mat = tmat,
-#'                                     n = n_samples,
-#'                                     point_estimate = FALSE)
+#'                                     n = n_samples)
 #'
 #' ### Utility
 #' utilitymod <- create_StateVals(utility_tbl, n = n_samples)