+++ R/params.R
@@ -37,24 +37,24 @@ check_params_joined <- function(x, inner_class, model_list){
   return(x)
 }
 
-create_params_list <- function(object, n, point_estimate, inner_class, new_class,
+create_params_list <- function(object, n, uncertainty, inner_class, new_class,
                                ...){
   n_objects <- length(object)
   params_list <- vector(mode = "list", length = n_objects)
   names(params_list) <- names(object)
   for (i in 1:n_objects){
-    params_list[[i]] <- create_params(object[[i]], n, point_estimate, ...)
+    params_list[[i]] <- create_params(object[[i]], n, uncertainty, ...)
   }
   return(new_params_list(params_list, inner_class = inner_class,
                          new_class = new_class))
 }
 
-create_params_joined <- function(object, n, point_estimate, inner_class){
+create_params_joined <- function(object, n, uncertainty, inner_class){
   n_models <- length(object$models)
   models <- vector(mode = "list", length = n_models)
   names(models) <- names(object$models)
   for (i in 1:n_models){
-    models[[i]] <- create_params(object$models[[i]], n, point_estimate)
+    models[[i]] <- create_params(object$models[[i]], n = n, uncertainty = uncertainty)
   }
   return(new_params_joined(models, times = object$times, inner_class = inner_class))
 }
@@ -62,15 +62,14 @@ create_params_joined <- function(object, n, point_estimate, inner_class){
 #' Create a parameter object from a fitted model
 #' 
 #' `create_params` is a generic function for creating an object containing 
-#' parameters from a fitted statistical model. If `point_estimate = FALSE`,
+#' parameters from a fitted statistical model. If `uncertainty != "none"`,
 #' then random samples from suitable probability distributions are returned.
 #' @param object A statistical model to randomly sample parameters from.  
-#' @param n Number of random observations to draw. 
-#' @param point_estimate If TRUE, then the point estimates are returned and
-#' and no samples are drawn.
-#' @param bootstrap If `bootstrap` is `FALSE` or not specified, then `n` parameter sets are 
-#' drawn by sampling from a multivariate normal distribution. If `bootstrap` is `TRUE`, then 
-#' parameters are bootstrapped using [bootstrap]. 
+#' @param n Number of random observations to draw. Not used if `uncertainty = "none"`.
+#' @param uncertainty Method determining how parameter uncertainty should be handled. 
+#' If `"normal"`, then parameters are randomly drawn from their multivariate normal
+#' distribution. If `"bootstrap`, then parameters are bootstrapped using [`bootstrap`].
+#' If `"none`, then only point estimates are returned.
 #' @param max_errors Equivalent to the `max_errors` argument in [bootstrap]. 
 #' @param ... Further arguments passed to or from other methods. Currently unused.
 #' @return An object prefixed by `params_`. Mapping between `create_params` 
@@ -194,8 +193,13 @@ check.params_lm_list <- function(object){
 
 #' @export
 #' @rdname create_params
-create_params.lm <- function(object, n = 1000, point_estimate = FALSE, ...){
-  if (point_estimate == FALSE){
+create_params.lm <- function(object, n = 1000, uncertainty = c("normal", "none"),
+                             ...){
+  uncertainty <- deprecate_point_estimate(list(...)$point_estimate, uncertainty,
+                                          missing(uncertainty))
+  uncertainty <- match.arg(uncertainty)
+  
+  if (uncertainty == "normal"){
     coefs_sim <- MASS::mvrnorm(n, stats::coef(object), stats::vcov(object))
     if(is.vector(coefs_sim)) coefs_sim <- matrix(coefs_sim, nrow = 1)
     return(new_params_lm(coefs = coefs_sim,
@@ -213,8 +217,9 @@ create_params.lm <- function(object, n = 1000, point_estimate = FALSE, ...){
 #' @export
 # #' @rdname create_params
 #' @keywords internal
-create_params.lm_list <- function(object, n = 1000, point_estimate = FALSE, ...){
-  return(create_params_list(object, n, point_estimate, 
+create_params.lm_list <- function(object, n = 1000, uncertainty = c("normal", "none"),
+                                   ...){
+  return(create_params_list(object, n = n, uncertainty = uncertainty, 
                             inner_class = "params_lm", new_class = "params_lm_list"))
 }
 
@@ -472,8 +477,12 @@ flexsurvreg_aux <- function(object){
 
 #' @export
 #' @rdname create_params
-create_params.flexsurvreg <- function(object, n = 1000, point_estimate = FALSE, ...){
-  if (point_estimate == FALSE){
+create_params.flexsurvreg <- function(object, n = 1000, uncertainty = c("normal", "none"),
+                                      ...){
+  uncertainty <- deprecate_point_estimate(list(...)$point_estimate, uncertainty,
+                                          missing(uncertainty))
+  uncertainty <- match.arg(uncertainty)
+  if (uncertainty == "normal"){
     sim <- flexsurv::normboot.flexsurvreg(object, B = n, raw = TRUE, 
                                           transform = TRUE)
     n_samples <- n
@@ -551,10 +560,10 @@ params_mlogit_list <- function(...){
                                new_class = "params_mlogit_list")))
 }
 
-create_coef_multinom <- function(object, n = 1000, point_estimate = FALSE, ...){
+create_coef_multinom <- function(object, n = 1000, uncertainty){
   # Extract/simulate coefficients
   coefs <- c(t(stats::coef(object)))
-  if (point_estimate == FALSE){
+  if (uncertainty == "normal"){
     coefs_sim <- MASS::mvrnorm(n = n, 
                                mu = coefs,
                                Sigma = stats::vcov(object))
@@ -594,9 +603,13 @@ create_coef_multinom <- function(object, n = 1000, point_estimate = FALSE, ...){
 
 #' @export
 #' @rdname create_params
-create_params.multinom <- function(object, n = 1000, point_estimate = FALSE, ...){
-  coefs <- create_coef_multinom(object, n, point_estimate, ...)
-  if (point_estimate){
+create_params.multinom <- function(object, n = 1000, uncertainty = c("normal", "none"),
+                                   ...){
+  uncertainty <- deprecate_point_estimate(list(...)$point_estimate, uncertainty,
+                                          missing(uncertainty))
+  uncertainty <- match.arg(uncertainty)
+  coefs <- create_coef_multinom(object, n, uncertainty)
+  if (uncertainty == "none"){
     n_samples <- 1
   } else{
     n_samples <- n
@@ -607,8 +620,8 @@ create_params.multinom <- function(object, n = 1000, point_estimate = FALSE, ...
 
 #' @export
 #' @rdname create_params
-create_params.multinom_list <- function(object, n = 1000, point_estimate = FALSE, ...){
-  return(create_params_list(object, n, point_estimate, 
+create_params.multinom_list <- function(object, n = 1000, uncertainty = c("normal", "none"), ...){
+  return(create_params_list(object, n = n, uncertainty = uncertainty, 
                             inner_class = "params_mlogit", new_class = "params_mlogit_list",
                             ...))
 }
@@ -644,24 +657,22 @@ check.params_surv_list <- function(object){
 
 #' @export
 #' @rdname create_params
-create_params.flexsurvreg_list <- function(object, n = 1000, point_estimate = FALSE, ...){
-  return(create_params_list(object, n, point_estimate, 
+create_params.flexsurvreg_list <- function(object, n = 1000, uncertainty = c("normal", "none"),
+                                           ...){
+  return(create_params_list(object, n = n, uncertainty = uncertainty, 
                             inner_class = "params_surv", new_class = "params_surv_list"))
 }
 
 #' @export
 #' @rdname create_params
-create_params.partsurvfit <- function(object, n = 1000, point_estimate = FALSE, 
-                                      bootstrap = TRUE, max_errors = 0, ...){
-  if (point_estimate == TRUE & bootstrap == TRUE){
-    msg <- paste0("When 'point_estimate' = TRUE and 'bootstrap' = TRUE, the 'point_estimate' ",
-                  "argument is ignored and bootstrap replications are generated.")
-    warning(msg, call. = FALSE)
-  }
-  if(bootstrap){
+create_params.partsurvfit <- function(object, n = 1000, 
+                                      uncertainty = c("normal", "bootstrap", "none"), 
+                                      max_errors = 0, ...){
+  uncertainty <- match.arg(uncertainty)
+  if(uncertainty == "bootstrap"){
     res <- bootstrap(object, B = n, max_errors = max_errors)
   } else{
-    res <- create_params(object$models, n = n, point_estimate = point_estimate)
+    res <- create_params(object$models, n = n, uncertainty = uncertainty)
   }
   return(res)
 }
@@ -751,14 +762,18 @@ check.params_joined_surv_list <- function(object, inner_class){
 #' @export
 # #' @rdname create_params
 #' @keywords internal
-create_params.joined_flexsurvreg <- function(object, n = 1000, point_estimate = FALSE, ...){
-  return(create_params_joined(object, n, point_estimate, "params_surv"))
+create_params.joined_flexsurvreg <- function(object, n = 1000, uncertainty = c("normal", "none"),
+                                             ...){
+  return(create_params_joined(object, n = n, uncertainty = uncertainty,
+                              inner_class =  "params_surv"))
 }
 
 #' @export
 # #' @rdname create_params
 #' @keywords internal
-create_params.joined_flexsurvreg_list <- function(object, n = 1000, point_estimate = FALSE, ...){
-  return(create_params_joined(object, n, point_estimate, "params_surv_list"))
+create_params.joined_flexsurvreg_list <- function(object, n = 1000, uncertainty = c("normal", "none"),
+                                                  ...){
+  return(create_params_joined(object, n = n, uncertainty = uncertainty, 
+                              inner_class = "params_surv_list"))
 }